+++ src/redux/reservations/reservationsSlice.js
@@ -6,7 +6,7 @@ const API_BASE = 'https://doctors-appointment-0mkx.onrender.com/api/v1/';
 // Helper function to get the JWT token from localStorage
 const gettoken = () => {
   const jwtToken = localStorage.getItem('jwtToken');
-  console.log(`token is ${jwtToken}`);
+
   return jwtToken;
 };
 
@@ -21,29 +21,40 @@ const axiosInstance = axios.create({
 export const createReserve = createAsyncThunk(
   'reserve/createReserve',
   async (payload) => {
-    const details = {
+    const ree = {
       appointment_date: payload.date,
-      appointment_duration: payload.duration,
+      appointment_duration: payload.time,
       doctor_id: payload.doctor,
       city: payload.city,
     };
-    const response = await axiosInstance.post('/appointments', details);
-    return response.data;
+
+    try {
+      const response = await axiosInstance.post('/appointments', ree);
+
+      return response.data;
+    } catch (error) {
+      throw new Error('Error occured try again');
+    }
   },
 );
 
-export const getReservations = createAsyncThunk('reserve/getReserve', async () => {
-  const response = await axiosInstance.get('/appointments');
-  return response.data;
+export const getReserve = createAsyncThunk('reserve/getReserve', async () => {
+  try {
+    const response = await axiosInstance.get('/appointments');
+
+    return response.data;
+  } catch (error) {
+    throw new Error('Error occured try again');
+  }
 });
 
-export const deleteReserve = createAsyncThunk(
-  'reserve/deleteReserve',
-  async (payload) => {
-    await axiosInstance.delete(`/appointments/${payload}`);
-    return payload;
-  },
-);
+// export const deleteReserve = createAsyncThunk(
+//   'reserve/deleteReserve',
+//   async (payload) => {
+//     await axiosInstance.delete(`/appointments/${payload}`);
+//     return payload;
+//   },
+// );
 
 const initialState = {
   reservations: [],
@@ -61,7 +72,7 @@ const ReservationSlice = createSlice({
         localStorage.setItem('success', JSON.stringify(action.payload));
       }
     });
-    builder.addCase(getReservations.fulfilled, (state, action) => {
+    builder.addCase(getReserve.fulfilled, (state, action) => {
       state.reservations = action.payload;
     });
   },
